Skip pip's default caching behavior
This caching includes the file that was downloaded over http,
as well as this new wheel cache. Neither of which are valuable
and just add bloat into the resulting images. Similar to removing
the apt caches on disk at each layer. This also makes pip install's
a tiny bit faster since pip avoids trying to automatically create
wheels (which is fast itself, but again, unnecessary in this world).