fix certifi certificates with OpenSSL<1.0.2
The certifi bundled certs are all 2048bit by default, but
OpenSSL<1.0.2 has problems loading those sometimes. Since we're
currently using the default python docker image, which uses debian, 
which won't allow easy installation of a newer OpenSSL, we need to
work around that problem.

This patch activates the compatibility mode with 1024bit certs of 
certifi for requests, if certifi and OpenSSL<1.0.2 are installed.